added base unit quantities to purchase and inventories
The quantity of a food item in a purchase or inventory is not calculated
in base units and then stored along side the string in the database.
A before_save handler is used to keep this information up to date.  As
of this commit, the column can be changed directly via outside
influence.